+++ tests/commands/exporter_option_git_ref.rs
@@ -104,7 +104,7 @@ fn export_using_option_git_ref_with_value_root_commit_sha1() -> Result<()> {
     let output = "+ export file : nominal.drawio
 - export page 1 : Page-1
 \\ generate pdf file
-- export page 2 : Page 2
+- export page 2 : Page-2
 \\ generate pdf file";
 
     drawio_exporter